chore: move creators and thunks into separate files
