+++ ui/src/shared/components/GetResources.tsx
@@ -7,7 +7,7 @@ import {getLabels} from 'src/labels/actions'
 import {getBuckets} from 'src/buckets/actions/thunks'
 import {getTelegrafs} from 'src/telegrafs/actions/thunks'
 import {getPlugins} from 'src/dataLoaders/actions/telegrafEditor'
-import {getVariables} from 'src/variables/actions'
+import {getVariables} from 'src/variables/actions/thunks'
 import {getScrapers} from 'src/scrapers/actions/thunks'
 import {getDashboardsAsync} from 'src/dashboards/actions'
 import {getTasks} from 'src/tasks/actions/thunks'
