+++ ui/src/variables/components/UpdateVariableOverlay.tsx
@@ -17,7 +17,7 @@ import {
 import VariableArgumentsEditor from 'src/variables/components/VariableArgumentsEditor'
 
 // Actions
-import {updateVariable} from 'src/variables/actions'
+import {updateVariable} from 'src/variables/actions/thunks'
 
 // Utils
 import {extractVariablesList} from 'src/variables/selectors'
